Simple Product validations
Presence was an issue so I validated for presence. The main problem
that this commit solves is the numericality validation of the price
attribute. I check for the value to be 'greater than or equal to'
0.01 and not 0 because 0.01 is more strict.

If a user entered 0.001 as a price, active record would round to
two decimal places and the value would get validated/added to the
persistence layer.