Cart Retrieval and Creation
Create a CurrentCart concern than can be used by all controllers to
deal with the Cart. This handles the happy path of a cart already existing
when a user needs it, and the unhappy path of there not being a cart when
for the user when he/she needs one.

The set_cart method retrieves the appropriate cart
from the database, using the session to do so. If the session does not
contain the information about the cart, Then a new cart is created and
the session is assigned information about that new cart.